Consolidate ContextSanityTest into OtelContextStorageModeTest
Move testMaxValueContext, testSequentialContextUpdates,
testRepeatedContextWrites, testNestedContextUpdates, testThreadIsolation
into OtelContextStorageModeTest. Drop tests fully covered by existing
ones (testBasicContextWriteRead, testZeroContext, testVariedContextValues).
Delete ContextSanityTest.java.
